feat: light sensing and led fading
Add support for a light sensor on one of the GPIO pins. The ADC
will read its value periodically and keep a moving average, used
later to control the brightness of the main clock LEDs.

Smooth out changes in LED brightness (both due to the light sensor
and due to LEDs being turned on or off) to improve aesthetics.